Refactor the test folder and add license headers
